more replicaset and replicaset's methods to replicaset.go
